refactor: folder naming refactoring
