+++ src/App.tsx
@@ -2,7 +2,7 @@ import { useForm } from 'react-hook-form'
 
 import { Button } from '@/components/ui/button'
 import { FlexContainer } from '@/shared/ui/flex-container'
-import { ControlledCheckbox } from '@/shared/ui/from-components/controlled-checkbox'
+import { ControlledCheckbox } from '@/shared/ui/form-components/controlled-checkbox'
 import { Layout } from '@/shared/ui/layout'
 
 import { Header } from './shared/ui/header'
